[REFACTOR] Google 소셜 로그인과 공유를 위한 클래스 파일 명 & 변수 명 수정 #113
