+++ src/main/java/indipage/org/indipage/auth/service/PublicKeyGenerator.java
@@ -1,9 +1,7 @@
-package indipage.org.indipage.auth.service.apple;
+package indipage.org.indipage.auth.service;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-import org.springframework.util.Base64Utils;
-
+import indipage.org.indipage.auth.service.JWKs;
+import indipage.org.indipage.auth.service.JWKs.JWKey;
 import java.math.BigInteger;
 import java.security.KeyFactory;
 import java.security.NoSuchAlgorithmException;
@@ -11,29 +9,32 @@ import java.security.PublicKey;
 import java.security.spec.InvalidKeySpecException;
 import java.security.spec.RSAPublicKeySpec;
 import java.util.Map;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+import org.springframework.util.Base64Utils;
 
 @Component
 @RequiredArgsConstructor
-public class ApplePublicKeyGenerator {
+public class PublicKeyGenerator {
 
     private static final String KEY_OF_HEADER_ALGORITHM = "alg";
     private static final String KEY_OF_HEADER_KEY_ID = "kid";
     private static final int POSITIVE_SIGN = 1;
 
-    public PublicKey generatePublicKey(final Map<String, String> headers, final ApplePublicKeys publicKeys) {
-        ApplePublicKeys.ApplePublicKey publicKey = getApplePublicKeyWithDecodedHeader(headers, publicKeys);
-        RSAPublicKeySpec publicKeySpec = generateRSAPublicKeySpecFromApplePublicKey(publicKey);
+    public PublicKey generatePublicKey(final Map<String, String> headers, final JWKs publicKeys) {
+        JWKey publicKey = getRightJWKWithDecodedHeader(headers, publicKeys);
+        RSAPublicKeySpec publicKeySpec = generateRSAPublicKeySpecFromJWK(publicKey);
         return generatePublicKeyFromKeySpecAndKeyType(publicKey.getKty(), publicKeySpec);
     }
 
-    private ApplePublicKeys.ApplePublicKey getApplePublicKeyWithDecodedHeader(final Map<String, String> headers, final ApplePublicKeys publicKeys) {
+    private JWKey getRightJWKWithDecodedHeader(final Map<String, String> headers, final JWKs publicKeys) {
         String alg = headers.get(KEY_OF_HEADER_ALGORITHM);
         String kid = headers.get(KEY_OF_HEADER_KEY_ID);
 
-        return publicKeys.getPublicKey(alg, kid);
+        return publicKeys.getRightJWK(alg, kid);
     }
 
-    private RSAPublicKeySpec generateRSAPublicKeySpecFromApplePublicKey(final ApplePublicKeys.ApplePublicKey publicKey) {
+    private RSAPublicKeySpec generateRSAPublicKeySpecFromJWK(final JWKey publicKey) {
         byte[] nBytes = Base64Utils.decodeFromUrlSafeString(publicKey.getN());
         byte[] eBytes = Base64Utils.decodeFromUrlSafeString(publicKey.getE());
