+++ src/main/java/indipage/org/indipage/auth/service/apple/AppleOAuthClient.java
@@ -1,7 +1,9 @@
 package indipage.org.indipage.auth.service.apple;
 
 import indipage.org.indipage.auth.dto.OAuthUserResponseDto;
+import indipage.org.indipage.auth.service.JWKs;
 import indipage.org.indipage.auth.service.OAuthClient;
+import indipage.org.indipage.auth.service.PublicKeyGenerator;
 import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import org.springframework.cloud.openfeign.FeignClient;
@@ -17,14 +19,14 @@ public class AppleOAuthClient implements OAuthClient {
 
     private final ApplePublicKeyClient publicKeyClient;
     private final AppleIdentityTokenProcessor identityTokenProcessor;
-    private final ApplePublicKeyGenerator applePublicKeyGenerator;
+    private final PublicKeyGenerator publicKeyGenerator;
 
     @Override
     public OAuthUserResponseDto getUser(String accessToken) {
         Map<String, String> header = identityTokenProcessor.getParsedHeader(accessToken);
-        ApplePublicKeys applePublicKeys = publicKeyClient.getApplePublicKeys();
+        JWKs JWKs = publicKeyClient.getJWKs();
 
-        PublicKey publicKey = applePublicKeyGenerator.generatePublicKey(header, applePublicKeys);
+        PublicKey publicKey = publicKeyGenerator.generatePublicKey(header, JWKs);
 
         Claims claims = identityTokenProcessor.extractClaimsFromIdentityToken(accessToken, publicKey);
 
@@ -36,6 +38,6 @@ public class AppleOAuthClient implements OAuthClient {
     @FeignClient(name = "apple-public-key-client", url = "https://appleid.apple.com/auth")
     public interface ApplePublicKeyClient {
         @GetMapping("/keys")
-        ApplePublicKeys getApplePublicKeys();
+        JWKs getJWKs();
     }
 }
