+++ src/main/java/indipage/org/indipage/auth/service/google/GoogleIdTokenProcessor.java
@@ -1,0 +1,2 @@
+package indipage.org.indipage.auth.service.google;public class GoogleIdTokenProcessor {
+}
