+++ src/main/java/indipage/org/indipage/auth/service/google/GoogleOAuthClient.java
@@ -1,0 +1,2 @@
+package indipage.org.indipage.auth.service.google;public class GoogleOAuthClient {
+}
